+++ web/src/components/forms/JiraDestinationForm/index.tsx
@@ -16,4 +16,4 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-export { default } from './JiraDestination';
+export { default } from './JiraDestinationForm';
